resolve issue with outputting double as int; unaware it was meant to be double

+++ Agent.h
@@ -72,7 +72,7 @@ protected:
 private:
     
     int health;
-    int speed;
+    double speed;
     
     
     enum class Agent_state_e {
